pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "cloud-sre-env"
version = "0.1.0"
description = "Cloud SRE DevOps environment for Meta Hackathon"
dependencies = [
    "openenv-core>=0.1.0",
    "fastapi>=0.111.0",
    "pydantic>=2.7.0",
    "uvicorn>=0.30.1",
    "openai>=1.30.0",
    "httpx>=0.27.0",
    "websocket-client>=1.9.0",
    "websockets>=16.0",
]

[project.scripts]
server = "server.app:main"
